+++ packages/message/src/Granules.ts
@@ -10,6 +10,7 @@
 
 import get from 'lodash/get';
 import { Message } from '@cumulus/types';
+import pick from 'lodash/pick';
 
 /**
  * Get granules from execution message.
@@ -23,3 +24,19 @@ import { Message } from '@cumulus/types';
 export const getMessageGranules = (
   message: Message.CumulusMessage
 ): unknown[] | undefined => get(message, 'payload.granules');
+
+/**
+ * Get selective cnm fields from execution message
+ *
+ * @param {Message.CumulusMessage} message - An execution message
+ * @returns {Object|undefined} object with cnm message fields
+ *
+ * @alias module:Granules
+ */
+export const getMessageCnm = (
+  message: Message.CumulusMessage
+): object | undefined => {
+  const fields = ['identifier', 'submissionTime', 'receivedTime', 'processCompleteTime'];
+  const cnm = get(message, 'meta.cnmResponse') || get(message, 'meta.cnm');
+  return cnm ? pick(cnm, fields) : cnm;
+};